+++ API/ComplexFieldInterface.php
@@ -10,9 +10,9 @@ interface ComplexFieldInterface
     /**
      * Return a non primitive field value - or a primitive field value with custom transformation
      *
-     * @param $fieldValueArray The definition of the field value, as used in the yml file
+     * @param $fieldValue The definition of the field value, as used in the yml file. Can be an array or a scalar value...
      * @param array $context The context for execution of the current migrations. Contains f.e. the path to the migration
      * @return mixed the object / array / scalar which will be passed to the setField() call of a content update or creation structure
      */
-    public function createValue($fieldValueArray, array $context = array());
+    public function createValue($fieldValue, array $context = array());
 }