+++ Core/ComplexField/EzBinaryFile.php
@@ -8,20 +8,40 @@ use Kaliop\eZMigrationBundle\API\ComplexFieldInterface;
 class EzBinaryFile extends AbstractComplexField implements ComplexFieldInterface
 {
     /**
-     * @param array $fieldValueArray The definition of teh field value, structured in the yml file
+     * @param array|string $fieldValue The path to the file or an array with 'path' key
      * @param array $context The context for execution of the current migrations. Contains f.e. the path to the migration
      * @return BinaryFileValue
      */
-    public function createValue($fieldValueArray, array $context = array())
+    public function createValue($fieldValue, array $context = array())
     {
-        $filePath = dirname($context['path']) .  '/files/' . $fieldValueArray['path'];
+        $mimeType = '';
+        $fileName = '';
+        if (is_string($fieldValue)) {
+            $filePath = $fieldValue;
+        } else {
+            $filePath = $fieldValue['path'];
+            if (isset($fieldValue['filename'])) {
+                $fileName = $fieldValue['filename'];
+            }
+            if (isset($fieldValue['mime_type'])) {
+                $mimeType = $fieldValue['mime_type'];
+            }
+        }
+
+        // default format: path is relative to the 'files' dir
+        $realFilePath = dirname($context['path']) .  '/files/' . $filePath;
+
+        // but in the past, when using a string, this worked as well as an absolute path, so we have to support it as well
+        if (!is_file($realFilePath) && is_file($filePath)) {
+            $realFilePath = $filePath;
+        }
 
         return new BinaryFileValue(
             array(
-                'path' => $filePath,
-                'fileSize' => filesize($filePath),
-                'fileName' => basename($filePath),
-                'mimeType' => mime_content_type($filePath)
+                'path' => $realFilePath,
+                'fileSize' => filesize($realFilePath),
+                'fileName' => $fileName != '' ? $fileName : basename($realFilePath),
+                'mimeType' => $mimeType != '' ? $mimeType : mime_content_type($realFilePath)
             )
         );
     }